pyproject.toml
[project]
name = "pdf2text"
version = "0.1.0"
description = ""
authors = [
    {name = "coela-oss"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "invoke (>=2.2.0,<3.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pdfminer-six (>=20240706,<20240707)",
    "polib (>=1.2.0,<2.0.0)",
    "babel (>=2.17.0,<3.0.0)"
]
package-mode = false

[tool.poetry.scripts]
pdf2text = "cli.main:entry"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
